feat: using Jest's `shard` option for test load balancing
We can use jest's `shard` option to implement test load balancing. This required updating to jest v28, as sharding was introduced in this version.
For now the shard count is chosen manually and is statically set, however, this may be optimised in the future.
Test load balancing is now utilised for both `test` and `build` (platform) pipelines.